Updated openstack/openstack Project: openstack-infra/subunit2sql  3bd7f9ff3f2e7c653fb9dc956dae1199777a0f6c
Add min ver to extra libs for graph

This commit adds min version to extra libs for graph command. When we
introduced a new graph type[1], we noticed that the matplotlib's min
version is required to use "style"[2]. However, regarding pandas, I'm
not confident with it. But an error occurred in my environment, when I
set pandas==0.10. (I suppose it's depends on OS or somethings.)

[1] Change-Id: I369717c84df0217a9a0cf17b7f9e975e580ce5a7
[2] http://matplotlib.org/users/whats_new.html#style-package-added

Change-Id: Iedb5203fb71ea522454b3f50a923fe8fc1ee1a74