Updated openstack/openstack Project: openstack/horizon  073ce8c5fd6ed121bf7d27f2234eded064547261
Magic Search now inherits from Bootstrap Theme

The magic search css was not themeable out of the box.  Also, we were
overriding almost every style in it with our own.  This causes
unneccessary complexity in the css; we should just not include it and
provide css that uses theme variables.

Partially-Implements: blueprint horizon-theme-css-reorg

Change-Id: I116e32b64bd6ba4c8f3e65fd1e145387f3071663