Updated openstack/openstack Project: openstack/horizon  6bfb80dc5a753e57391680ea092850afd823608d
Transfer table should update allocatedIds on allocated change

Transfer table does a one time allocation
for allocatedIds and does not watch changes to
allocated. As a result, when the allocated
list is fetched at a later point, the allocatedIds
are not updated resulting in a broken view.

Partially-Implements: blueprint ng-flavors
Closes-Bug: #1531334

Change-Id: I807c682b69081f90fc3d74675e8ca149bb855521