Updated openstack/openstack Project: openstack/nova  34cf19c8a4dababd1e723c0364a80823c0ac2637
Add 'hw:cpu_thread_policy=require' scheduling

The 'require' CPU threads policy ensure that hosts are required to have
an SMT architecture and that it be possible to schedule an instance
utilising the thread siblings.

The 'require' policy should and will only be used when explicitly
requested by the instance or image. The existing 'prefer' policy
(though it wasn't called this before) is retained and will be used both
when requested and as the default, per the spec. A stub is included for
the remaining 'isolate' case.

It is necessary to add a filter to exclude hosts with no sibling
support. This ensures that hosts that could not implement this
threading policy are quickly discarded.

Finally, some tests that tested the 'require'-like aspects of the
'prefer' policy are reworked to request this policy explicitly.

Change-Id: Idedf16a734d9d85db08269d05f8c147959a17679
Implements: blueprint virt-driver-cpu-thread-pinning