Updated openstack/openstack Project: openstack/swift  f7c8a1834e7e8faf580e1bc82d8646cdd5909b02
Functional tests for if-match with multiple etags

Multiple etags can be provided on an if-match or if-none-match
request. This is currently being tested in the unit tests, but not
in the functional tests. Since these etags can be modified by
middleware, we need functional tests to assert multiple-etag
requests are handled correctly.

Change-Id: Idc409c85e8aa82b59dc2bc28af6ca2617de82699