Updated openstack/openstack Project: openstack-infra/project-config  87b7e1ab8151d4ecdda9d4c062e5f99ff50744fa
Constraints jobs non-voting for neutron-vpnaas

This takes the experimental constraints based jobs and makes them
non-voting for non-Kilo release, as the next step in the process of
making these jobs available for master.

There are several additional changes and notes on this patch. First,
the unused experimental job, grenade-dsvm-neutron-vpnaas, was
previously remove. This commit removes the corresponding template,
and the lines to make this job non-voting.

Second, for non-Kilo branches, both non-constraint and constraint
jobs will run. In the next commit, when the constraints based jobs
are made voting, the non-constraints based jobs will only be run
for Kilo.

Third, the neutron-vpnaas-docs-constraints job is added to the
post queue. Currently, there is no docs job in the post queue
(e.g. no openstack-server-publish-jobs) for neutron-vpnaas. In the
next commit, openstack-server-publish-jobs-constraints will be
added for non-Kilo releases.

Fourth, the non-constraints coverage job will only run for Kilo
releases (the new constraints based job will run in post queue
for newer releases).

Change-Id: Ib131ca1686a78650a2e4be768746a83ed13307ee
Partial-Bug: #1522503