Updated openstack/openstack Project: openstack-dev/hacking  7f496e2537befbad3f2f36a2633ad83f52494084
Do not call "rm -rf" on arbitrary user input

While adopting this for bandit in
I38bfc2f8298761ebf86152933630f629e545029d we noticed that $project might
be something we do not want to unconditionally remove forcefully and
recursively. This creates a temporary directory for the clone and
removes only that temporary directory at the end. This allows us to
avoid using "rm -rf" with user input.

Change-Id: I17dae5da0a6b88a95035468df271b48641353e5a