Updated openstack/openstack Project: openstack/keystonemiddleware  14cfc270982af415bf9fa7411694422b454a68d4
Fix tests to work with keystoneauth1 2.2.0

Three tests were failing due to commit bbd85fd in the 2.2.0 release of
keystoneauth1, which changed the way that the user agent string is
generated. auth_token only needs to verify its part of the user agent
string rather than the entire string (the latter part of which is
controlled by keystoneauth).

Change-Id: I1e859440e8619857c92d3fbcbdd3b9fd2dc04e1e