Updated openstack/openstack Project: openstack/horizon  9a7ba89e4935a920279e2b6f429a13e34e667ca5
Handle absence of cinder default volume type

The code previously did not set the default_vol_type attribute if the
fetch of the default volume type failed (because none is declared in
the service). This patch defaults the attribute to the expected None
value, allowing volume creation to continue.

Change-Id: Ic9e14ae4cb23023a270ad4aec4e1a830a11215dd
Fixes-Bug: 1518844