Updated openstack/openstack Project: openstack-infra/release-tools  16eae5860c523aa4aa63f7e6740c5940bbb0e2ea
default to sending mail directly via lists.openstack.org

When the announce script runs on a CI node, we need to ensure the email
is sent before letting the job finish and destroying the node. Rather
than using an MTA on localhost, connect directly to lists.openstack.org
and push the mail there. That will block until the email is sent, at
which point it is safe for the job to exit.

Change-Id: I60be241402d6cc42f4219edec846a3c085019058