Updated openstack/openstack Project: openstack/heat  198ee80524a189a465b1e787cbac6f2d3023c91b
Remove unconditional return of user_id

This removes a legacy backwards compatibility codepath. Pre-Icehouse
stacks that have resources which inherit from StackUser may require
manual cleanup after deletion, as their stack user may not be deleted.

This fixes issues with subusers using software config. Heat should not
attempt to delete the stack user if one had not been created before.
The code that this patch removes caused an ID to always be returned.
Additionally, it set the ID in the resource metadata whether a stack
user existed or not, making it impossible to tell if one existed only
from data within the resource.

Change-Id: I50804998fdce1f7de06906ae9a2f47317721a6f6
Closes-bug: #1531606