Updated openstack/openstack Project: openstack/nova  523e8ed5c0a5d9d0b004d7b3a19aba0086f23b63
Remove "or 'reserved'" condition from reserve_block_device_name

This condition was added in 66f1a34cabc26930db2a1afed5a0f5ced4d1c015
in grizzly to fix a regression with the compute 2.x rpc api.

As of 91b35b22e753a4eb0c22c004bb12586970a95e11 we require compute
rpc api >= 4.0 so the 2.x rpc api compat is no longer a concern.

We should remove this condition since it's confusing, especially
considering that the block_device_mappings table has no unique constraint
yet we try to treat them as unique in most cases.

Note that the v2 and v2.1 REST APIs also require that a volumeId is
provided when attaching a volume (which is the only API that uses this
method), so the volume_id should always be provided and we don't need to
worry about this optional case.

Apparently there was never a unit test added for this case, so there are
no tests to update.

Change-Id: I5cde374a2621b22b094d14e357a399a2fb784287
Closes-Bug: #1533871