Updated openstack/openstack Project: openstack/tempest  379a76acaa58748ce67e4941de17b2050cf24f92
Fix create_test_server for multiple create request

Currently create_test_server function assume that if
either 'min_count' or 'max_count' are present in kwargs then
request is for multiple create.

But that is not valid assumption. If 'min_count' and 'max_count'
are present and equal to 1 then it is single server create request.

Nova also set them default to 1 if they are not passed.
https://github.com/openstack/nova/blob/master/nova/compute/api.py#L1133

This commits checks if they are greater than 1 than only assume that
request is for multiple create server.

Change-Id: Ie4f45eb8481cd1aff55877cb7f8be327d548031e