Updated openstack/openstack Project: openstack/heat  7be1dd7e4fa9c0e1564b2a99d59cdd6716f5b9b4
Convergence: Pick resource from dead engine worker

When a engine worker crashes or is restarted, the resources being
provisioned in it remain in IN_PROGRESS state. Next stack update should
pick these resources and work on them. The implementation is to set the
status of resource as FAILED and re-trigger check_resource.

Change-Id: Ib7fd73eadd0127f8fae47881b59388b31131daf4
Closes-Bug: #1501161