Updated openstack/openstack Project: openstack/swift  8460ddd6077beca4b3acd0942f665e89c4e2b16e
Make object-auditor storage-policy-aware

Previously, the object-auditor would always use a (replication)
DiskFileManager when walking through AuditLocations, which would cause
it to skip EC fragment archives with a warning like:

    Unexpected file <hash_path>/1452557187.03610#3.data:
    Invalid Timestamp value in filename '1452557187.03610#3.data'

Now, the AuditLocation's policy will be used to find an appropriate
manager to get the diskfile. Additionally, several .commit()s were added
to the auditor tests so the .durable will be written, which is required
when auditing EC fragment archives.

Change-Id: I960e7d696fd9ad704ca1872b4ac821f9078838c7
Closes-Bug: 1533002