Updated openstack/openstack Project: openstack-infra/project-config  4cc41b1dda6729f94c71856af6e9c6dc70e47be5
Identity API v3 only non-voting job in Tempest

There is an experimental gate job for Identity API v3 only for Tempest
repository. The same job is on the main check pipeline for DevStack.

The reason it hadn't been promoted to the Tempest main check pipeline
yet was that there were too many gates running there already. Now
after Juno EOL date it should be fine to acept it.

That job aims to verify that Tempest tests can run with the described
configuration, and consequently services are compatible with the
Identity v3 auth API.

This change moves the job from the experimental pipeline to the main
check pipeline, keeping it as non-voting for now.

The goal is, once there is enough confidence on its stability, to make
it voting. Currently it has failed 10% of the runs on periodic-qa [1].

[1] http://status.openstack.org/openstack-health/#/g/build_queue/periodic-qa

Change-Id: Iab052186b7da036b6e6b6ad52acdd1647e94494f