Updated openstack/openstack Project: openstack-dev/pbr  12940d5b9d93297290660cf4273d173577309cc7
Split changelog on nulls instead of (

The logic that splits the git log lines to find tags gets confused by
parentheses in the commit message and truncates the line at the first
one. To mitigate this, have git emit a null character and split on that.

As a consequence of doing and testing this, fix the bug where
non-semantic-version tags in the tag history would show up in the
ChangeLog.

Change-Id: I5e522d6d374534f2921aaded42780df97c3d1987