Updated openstack/openstack Project: openstack/manila  70a83985e07578ff20c39318d8d6d16ef56ff8c1
Fix 'extend' API for 2.7+ microversions

Extend/Shrink share is not working in microversion 2.7+.
When trying to extend/shrink a share, manila raises an
error telling that the new share size must be an integer.
Bug was due to improper use of get method on dictionary.
Corrected by using get method properly.

Change-Id: I700581d815da0fdd6addedc42f3e2ba528680e60
Closes-Bug: #1531536