Updated openstack/openstack Project: openstack/requirements  236706d358b26cfa7f046c1e64513d2b484d8789
Add google api python client

This is the Python client library for Google's discovery based APIs.
It will be used for backup operations with Google Cloud Storage
in openstack cinder google backup driver.

Is the library actively maintained?
  Yes(https://github.com/google/google-api-python-client).
Is the library good code?
  Yes.
Is the library python 3 compatible?
  Yes.
Is the library license compatible?
  Yes.
Is the library already packaged in the distros we target?
  Use pip to manage installation(https://developers.google.com/
  api-client-library/python/start/installation).
Is the function of this library already covered by other
libraries in global-requirements.txt?
  No.

Change-Id: I04f2e3dee1f5e001ad2f1cce15893b0965c082d7