Updated openstack/openstack Project: openstack-infra/storyboard  a91b276591c113abd4018c4a72fc95cdc883ddde
Work around task deletion making it impossible to get subscribers

When a task is deleted, we can't get the story it is in since the task
no longer exists in the database. This causes the function which gets
the list of affected subcribers to raise an exception.

This patch works around this by instead getting the list of affected
subscribers for the task's story instead. This is sub-optimal (people
subscribed to just the deleted task and not the other affected resources
won't get email, and people who are just subscribed to other tasks in
the story will get email they didn't ask for), but much easier in the
short term than stopping deletion.

In the long term, we should archive tasks and stories instead of
deleting them outright, which will make this problem go away.

Change-Id: I03ad6845772a1515708bef8020228a2abc26351f