Updated openstack/openstack Project: openstack-infra/release-tools  7710ae7105fc4d2d6418ce5e6b1bdae76b84d6cd
remove use of virtualenv from release scripts

No longer rely on the release scripts to have a virtualenv or to set one
up. All of the tools will be installed globally, and libraries from
system packages, when we move to running these scripts on the secure
nodes in the CI system.

Change-Id: I30d430de9a690878950a8f8775b33ca7ba2896be