Updated openstack/openstack Project: openstack-infra/project-config  7cbe842e6f15e7a716882acc2861e23e82643f04
Remove update_po from translation proposal script

So far we have updated some po files (files in a specific language) whenever
we downloaded files from our translation server. This was done
for python projects only but not for manuals and horizon - and
was broken for django_openstack_auth.

It was done to ensure that POT files (translation source) and PO files
(translated files in specific language) are in sync. But since we're
pushing to Zanata and downloading *all* translation files from there,
Zanata does the update step for us and we get all changes.

Remove update_po now, it's not needed.

Change-Id: I9a0dac230dbb8c47ae6d82aebbf7727c54f2d2cd