Updated openstack/openstack Project: openstack-infra/storyboard-webclient  17893038d2d5ac45b42200fbae3402655e9b2a52
Make task project titles link to projects

Clicking on the edit button will edit the task project. Clicking on
the task project title will now take the user to the project. Before,
when logged in, it was not possible to browse to a project from the
task view, only to edit it.

Change-Id: I73b5666c29c6d056235dc64f60004fca29437c8a