Updated openstack/openstack Project: openstack/nova  52a103a0bf59fb636108c1c6e407af58e7725d8f
Allow virt driver to define binding:host_id

This patch allows the virt driver to define the host_id field, with
instance.host as the base implementation. This is passed into the
networks API and sent to Neutron with the port-create calls.

Ironic ports will need to be created with an empty host_id field in
order to do late binding of the Neutron ports. For backwards
compatibility, the old behavior should be used when
node.network_provider is 'none'.

Implements: blueprint ironic-networks-support
Change-Id: I580dc9c5d3df5dcaf8a6933853609992be2d41f6