Updated openstack/openstack Project: openstack/nova  836b3705da4883ade4a9894f06cd4fb9d3561771
Fix undetected races when getting BDMs by volume id

This deprecates the BDM.get_by_volume_id() method and introduces two new ones.
The existing method will fail to detect that multiple instances can have BDMs
pointing at the same volume. This could happen via race, and will be normal
behavior when we get volume multi-attach support.

Instead of this method which takes instance as an optional parameter, a new
method is introduced which requires the property when searching for a single
BDM for a volume. Another method is added that allows us to fetch BDMs for
volumes where there should be only one instance attached. This method checks
that the single-owner consistency is correct, or fails otherwise.

Closes-Bug: #1533834
Co-Authored-By: Ildiko Vancsa <ildiko.vancsa@ericsson.com>
Co-Authored-By: Krisztian Gacsal <krisztian.gacsal@ericsson.com>
Change-Id: I06cdeef38aeb3f3e65dfc0ced8a75960b5d57f49