Updated openstack/openstack Project: openstack/horizon  0fb6bda0e47d5d4a6d1cf39bf74b5833d44d1397
Add the max value to the tooltip for the "Items per Page" field

Currently, the tooltip for the "Items per Page" field on the User
Settings page does not show the maximum value that the field could be.
Instead, the user must first enter the invalid value, only to receive an
error notification that then shows the max value that the field will
accept.

To resolve this issue, the maximum value is displayed along with the
other help text in the tooltip.  The maximum value has a default value
of 1000 unless set otherwise.

Change-Id: I267cb5ac067f7f0b27224ca8dbba0c8e19199477
Closes-Bug: #1418430