Updated openstack/openstack Project: openstack/nova  c0cd868ad2543831f7e675735aa2575e4917802e
Parse availability_zone in API

This moves the parsing of availability_zone, and as a consequence
force_host/force_node, into the API to prepare for building filter_properties
earlier.  This will be necessary in order to build and persist the request_spec
object before the instance has been saved to the db.

Change-Id: I55399a77fc8d474380f7c422dcef1e6d145a48c8