Updated openstack/openstack Project: openstack/nova  6538eb4419232e21765a04fe7746d9c8c9d19a90
libvirt: handle migrate_data as object in cleanup method

As of 69e01758076d0e89eddfe6945c8c7e423c862a49, the
migrate_data parameter is a LibvirtLiveMigrateData object
and doesn't have dict-compat so we have to handle it like an
object.

Also adds a unit test for this case which didn't exist before
which is probably why it wasn't caught in the previous change.

Change-Id: I5a676334d0082373b5c1895f060283cc942b2d40
Closes-Bug: #1534837