Updated openstack/openstack Project: openstack/neutron  7fd8d2d71f412fbe04bf7e04ff4b449f00e021d8
Move L2populationDbMixin to module-level functions

None of the L2populationDbMixin methods actually use 'self' for
anything. As the class is basically just used as a namespace and
modules already provide that, this patch gets rid of the mixin. This
makes the code simpler and easier to debug as inheritance doesn't buy
us anything in this case.

Change-Id: Ibf4dfe49a2ebc32d3909d3d7b579d2bb2ea3f61d