Updated openstack/openstack Project: openstack/neutron  92fb281257fb2be2b16c9ad4514afc36f4706fdd
Add opnfv tag to the list of auto-complete tags

The OPNFV project [1] is a reference implementation of the
ETSI NFV architecture built on open source components.
OpenStack, and Neutron in particular, are key elements of
this story, and many requirements and issues are driven and
submitted by OPNFV folks that work in both communities.

This tag makes sure that we can keep all of them together for
tracking purposes.

[1] http://superuser.openstack.org/articles/openstack-and-opnfv-strengthen-collaboration-for-telcos

Change-Id: Ie1675ef6f177558f579097fe035494b9380232d0