Updated openstack/openstack Project: openstack/neutron  d6d43b32ca825b6c3c2c908f5ff7bc50c736546e
Register RA and PD config options in l3-agent

In order for the l3-agent to see the RA and PD config options,
it needs to register them when it starts.  Noticed this when I
went to override something for a test and it wouldn't work.
It now passes the config down to radvd on start so the correct
values are picked-up.

Change-Id: Iec0e0d16eed4f12af77fcd4f0b93b641b1146293
Related-Bug: #1532338