Updated openstack/openstack Project: openstack/tempest  f558bb3302b232ebe0c2328ad58b25a0d0af99a5
Add different_host test on multiple nodes

Nova supports a scheduler_hint "different_host", but it was difficult
to test it because the gate didn't support multiple nodes.
Now the gate supports multiple nodes and we can test the scheduler_hint.
So this patch adds a different_host test.

Change-Id: Ib3365ac2783a0578c7a1a1e72d9b6c9cfea340f5
Related-Bug: #1521928
Related-Bug: #1526620